Moved buffer pool support to package: libs
