+++ libs/buffer_pool_for_uploading.go
@@ -1,23 +1,23 @@
-package goseaweedfs
+package libs
 
 import (
 	"bytes"
 	"sync"
 )
 
-var gBufferPool *BufferPool
-
-// !! For user who wants to use buffer pool when uploading, please explicitly call InitBufferPool(bufferLen, bufferCap int) !!
-func InitBufferPool(bufferLen, bufferCap int) *BufferPool {
-	gBufferPool = NewBufferPool(bufferLen, bufferCap)
-	buf := gBufferPool.Get() // pre-cache
-	gBufferPool.Put(buf)     // pre-cache
-	return gBufferPool
-}
-
-func GetBufferPool() *BufferPool {
-	return gBufferPool
-}
+//var gBufferPool *BufferPool
+//
+//// !! For user who wants to use buffer pool when uploading, please explicitly call InitBufferPool(bufferLen, bufferCap int) !!
+//func InitBufferPool(bufferLen, bufferCap int) *BufferPool {
+//	gBufferPool = NewBufferPool(bufferLen, bufferCap)
+//	buf := gBufferPool.Get() // pre-cache
+//	gBufferPool.Put(buf)     // pre-cache
+//	return gBufferPool
+//}
+//
+//func GetBufferPool() *BufferPool {
+//	return gBufferPool
+//}
 
 // BufferPool : will be used when uploading with multipart which will pre-allocate and reuse memory, and reduce memory usage significantly if we can estimate the file size we are uploading.
 type BufferPool struct {
